% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{assignFill}
\alias{assignFill}
\title{Assigns classification labels to pixels based on gene expression similarity.}
\usage{
assignFill(input, gene, df)
}
\arguments{
\item{input}{A list. Results from `spatialSimilarity()`. 
This includes the similarity table and classified pixel IDs.}

\item{gene}{Character. The name of the gene for which pixel classification is performed.}

\item{df}{Data frame containing pixel-level gene expression data.}
}
\value{
The input data frame with an additional column, `fill`, which assigns 
classification labels to pixels:
\itemize{
  \item{\strong{1}}{Pixels classified as similar (within the fold-change threshold).}
  \item{\strong{2}}{Pixels with greater expression in dataset X than Y}
  \item{\strong{3}}{Pixels with greater expression in dataset Y than X}
  \item{\strong{4}}{Pixels with gene expression below the threshold in both experiments.}
}
}
\description{
This function categorizes pixels into four groups based on their similarity 
in gene expression between two spatial experiments. The assigned classification 
is stored in a new column, `fill`, within the input data frame.
}
