% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageFunction.R
\name{getGenePixelDF}
\alias{getGenePixelDF}
\title{Extract Gene Expression Values for Shared Pixels Between Two SpatialExperiments}
\usage{
getGenePixelDF(x, y, gene, assayName = assayName)
}
\arguments{
\item{x}{A SpatialExperiment object representing the first dataset.}

\item{y}{A SpatialExperiment object representing the second dataset.}

\item{gene}{A character string specifying the gene of interest.}

\item{assayName}{A character string or numeric specifying the assay in the Spatial Experiment to use.}
}
\value{
A data frame with three columns:
\item{pixel}{Character vector of shared pixel IDs between both experiments.}
\item{y}{Numeric vector of gene expression values from `y` (second SpatialExperiment).}
\item{x}{Numeric vector of gene expression values from `x` (first SpatialExperiment).}
}
\description{
This function creates a data frame containing gene expression values for a 
specified gene across shared spatial pixels in two SpatialExperiment objects.
}
\examples{
getGenePixelDF(spatialExp1, spatialExp2, "Aqp4", 'pixelval')
}
