% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCorrelation.R
\name{matchingVariograms}
\alias{matchingVariograms}
\title{matchingVariograms}
\usage{
matchingVariograms(
  X.randomized,
  long,
  lat,
  delta,
  target_variog,
  prctile,
  ids,
  i
)
}
\arguments{
\item{X.randomized}{\code{numeric vector} random permutation of the values of
X across its x,y-coordinates.}

\item{long}{\code{numeric vector}: y-coordinates of points in dataset X}

\item{lat}{\code{numeric vector}: x-coordinates of points in dataset X}

\item{delta}{\code{numeric vector} Given a point in dataset X, the
percentage of neighbors which should be within the smoothing kernel. This
can be a list or a single numeric. Input to `locfit::lp` as `nn`}

\item{target_variog}{\code{variogram} variogram of dataset X that will be
used as target when doing the matching. See geoR::variog}

\item{prctile}{\code{numeric}: the max distance that should be used for
evaluating the fit between the variograms from the permuted data and the
target variogram. Can be determined from calculating all distances between
the points and then calculating a specified quantile. Input to
`geoR::variog` as `max.dist`}

\item{ids}{\code{integer}: indices from 1 to the number of points in the
dataset X}

\item{i}{\code{integer} the number indicating the ith permutation}
}
\value{
The output is returned as a list.
 \itemize{
  \item{\code{residus}}{numeric vector of length of delta, sum of squares of
  the residuals between the target variogram and the permuation variogram
  for each delta}
  \item{\code{delta.star.id}}{integer, the index of the delta that minimizes
  the residual sum of squares}
  \item{\code{hat.X.delta.star}}{numeric vector of length of X.randomized,
  permutation generated from optimal delta, whose variogram best matches
  the variogram of X}
  }
}
\description{
Function to create a permutation of spatial dataset X by
  starting with a randomly shuffled version of X and then smoothing it until
  the variogram of this permutation matches the variogram of the unmodified X
}
\examples{

data(quakes)

#remove duplicated positions
quakes_data <- quakes[!duplicated(cbind(quakes$lat, quakes$long)),]

lat <- quakes_data$lat
long <- quakes_data$long
X <- quakes_data$depth
ids <- 1:length(X)

#number of permutations
B <- 3

#sequence of deltas
delta <- seq(0.1,0.9,0.1)

# maximum distance for the variogram set at the 25\% percentile of
the distribution of pairs of distances:
dists <- dist(cbind(lat, long))
prctile <- quantile(dists, probs = 0.25)

# variogram of variable X that will be used as target when doing the matching:
target_variog <- geoR::variog(data = X, coords = cbind(long, lat), 
                              max.dist = prctile, option = "bin", 
                              messages = FALSE)

# create B random permutations of the values of X across locations:
X.randomized <- lapply(1:B, function(i) {
   sample(X, size = length(X), replace = FALSE)
})

# create B permutations of the values of X with variograms that match the 
#target variogram of X:
output <- lapply(1:B, function(i) {
  output <- matchingVariograms(X.randomized[[i]], long, lat, 
                               delta, target_variog, prctile, ids, i)
})

output

}
\references{
adapted from: 
Viladomat, Júlia et al. “Assessing the significance of global and
  local correlations under spatial autocorrelation: a nonparametric
  approach.” Biometrics vol. 70,2 (2014): 409-18. doi:10.1111/biom.12139
}
