% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCorrelation.R
\name{spatialCorrelationGeneExpWithinSample}
\alias{spatialCorrelationGeneExpWithinSample}
\title{spatialCorrelationGeneExpWithinSample}
\usage{
spatialCorrelationGeneExpWithinSample(
  input,
  nPermutations = 100,
  delta = NULL,
  maxDistPrctile = 0.25,
  returnPermutations = FALSE,
  assayName = NULL,
  nThreads = 1,
  BPPARAM = NULL,
  verbose = TRUE,
  seed = 0
)
}
\arguments{
\item{input}{\code{SpatialExperiment} A SpatialExperiment object. See
\code{assayName} parameter if the SpatialExperiment object has more than
one assay.}

\item{nPermutations}{\code{integer} or \code{double}: number of permutations
to generate to build the empirical null distribution. This number will
determine the precision of the p-value. Default is \code{100}, such that
the smallest possible p-value is \eqn{1 / 101}, about 0.0099}

\item{maxDistPrctile}{\code{numeric}: percentile of distances between pixels
to use as max distance in when calculating variograms. Default = 0.25. At
greater distances the variogram is less precise because there are fewer
pairs of points with that distance between them. Therefore, since the goal
is to minimize the difference between the variogram of X and those of its
permutations, the variogram should be subsetted to the percentile that is
more robust.}

\item{returnPermutations}{\code{logical}: indicate whether the dataframe
returned as output will have a column with the values of the permutations
used to calculate the null correlations and the empirical p-value. Default
is \code{FALSE}}

\item{assayName}{\code{character} or \code{integer} A character string or
numeric specifying the assay in the SpatialExperiment to use. Default is
\code{NULL}. If no value is supplied for \code{assayName}, then the first
assay is used as a default}

\item{nThreads}{\code{integer}: Number of threads for parallelization.
Default = 1. Inputting this argument when the \code{BPPARAM} argument is
\code{NULL} would set parallel execution back-end to be
\code{BiocParallel::MulticoreParam(workers = nThreads)}. We recommend
setting this argument to be the number of cores available
(\code{parallel::detectCores(logical = FALSE)}). If \code{BPPARAM} argument
is not \code{NULL}, the \code{BPPARAM} argument would override
\code{nThreads} argument.}

\item{BPPARAM}{\code{BiocParallelParam}: Optional additional argument for
parallelization. This argument is provided for advanced users of
\code{BiocParallel} for further flexibility for setting up
parallel-execution back-end. Default is NULL. If provided, this is assumed
to be an instance of \code{BiocParallelParam}.}

\item{verbose}{\code{logical}: indicate whether to print row number and name
to show progress as the function iterates through the rows of the
SpatialExperiments to calculate a correlation coefficient and empirical
p-value for each row}

\item{seed}{\code{integer}: Seed for the random number generator used to
generate noise in the variogram matching step. Ensures reproducibility of
empirical p-values regardless of parallelization back-end. Default is
\code{0}.}

\item{deltaX}{\code{list}: List of single numerics or list of numeric vectors
to use for delta, the parameter controlling the degree of smoothing in
permutations of X. The length of the list should the same as the number of
rows in the SpatialExperiment.  Delta is a proportion calculated by
dividing k neighbors by N total observations (columns) in X, where k is the
number of neighbors in the permutation of X that should be within the
radius smoothed by the Gaussian kernel to achieve the amount of
autocorrelation present in the original X. If a single delta is not known,
a sequence of deltas can be inputted and the best delta will be found such
that it minimizes the sum of squares of the residuals between the variogram
of the permutation generated from the delta and the variogram of the
target. Default is \code{NULL}. If no value is supplied for \code{deltaX},
\code{seq(0.1,0.9,0.1)}, the sequence of every 0.1 from 0.1 to 0.9, will be
used to find the best delta for each row (gene) in X.}

\item{deltaY}{\code{list}: List of single numerics or list of numeric vectors
to use for delta, the parameter controlling the degree of smoothing in
permutations of Y. \code{deltaY} is like \code{deltaX} but for permuting
data in Y instead of X. Default is \code{NULL}. If no value is supplied for
\code{deltaY}, \code{seq(0.1,0.9,0.1)}, the sequence of every 0.1 from 0.1
to 0.9, will be used to find the best delta for permutations for each row
(gene) in Y.}
}
\value{
The output is returned as a \code{data.frame}. The rownames are
  arbitrary. The names of the columns and their contents are as follows:
\itemize{
  \item{\code{correlationCoef}}{Pearson's correlation coefficient.}
  \item{\code{pValueNaive}}{the analytical p-value naively assuming independent
  observations}
  \item{\code{pValuePermuteX}}{the p-value when creating an empirical null from permutations
  of observations in X}
  \item{\code{pValuePermuteY}}{the p-value when creating an empirical null from
  permutations of observations in Y}
  \item{\code{deltaStarMedianX}}{the median delta star (the delta which
  minimizes the difference between the variogram of the permutation and the
  variogram of observations) across permutations of X}
  \item{\code{deltaStarMedianY}}{the median delta star across permutations of Y}
  \item{\code{deltaStarX}}{list of delta star for all permutations of X}
  \item{\code{deltaStarY}}{list of delta star for all permutations of Y}
  \item{\code{nullCorrelationsX}}{correlation coefficients for Y and all permuations of X}
  \item{\code{nullCorrelationsY}}{correlation coefficients for X and all permuations of Y}
  \item{\code{permutationsX}}{(optional) a N x B matrix, where N is the length of X and B is `nPermutations`.
  Each column is the resulting values of a permutation of X}
  \item{\code{first}}{the name of the first row in the pair}
  \item{\code{second}}{the name of the second row in the pair}
  }
}
\description{
Function to calculate Pearson's correlation between rows from
  one SpatialExperiment dataset. To replace the analytical p-value which
  results in a high false positive rate for autocorrelated spatial patterns,
  it calculates empirical p-values from empirical null distributions
  generated from permuting the data and then smoothing to maintain the
  original degree of autocorrelation
}
\examples{

data(speKidney)

##### Rasterize to get pixels at matched spatial locations #####
rastKidney <- SEraster::rasterizeGeneExpression(speKidney,
               assay_name = 'counts', resolution = 0.2, fun = "mean",
               BPPARAM = BiocParallel::MulticoreParam(), square = FALSE)

sc_within_sample <- spatialCorrelationGeneExpWithinSample(
                      input = rastKidney,
                      assayName = "A"
                      )
sc_within_sample

}
