% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageFunction.R
\name{threshold}
\alias{threshold}
\title{Threshold and preprocess gene expression data}
\usage{
threshold(genePixelDF, t1, t2)
}
\arguments{
\item{genePixelDF}{A data frame containing gene expression for one gene for
each pixel. It should have columns "x" and "y" representing expression
levels in two SpatialExperiments.}

\item{t1}{\code{numeric}: Threshold value for gene expression in the first
SpatialExperiment (x).}

\item{t2}{\code{numeric}:Threshold value for gene expression in the second
SpatialExperiment (y).}
}
\value{
A filtered and processed data frame where:
        - NA values are removed,
        - Pixels with both x and y values below their respective thresholds are excluded,
        - Zero values are approximated to 0.001 to avoid numerical instability.
}
\description{
This function processes a gene-to-pixel data frame by removing missing
values, applying a threshold to filter out low-expression pixels, and
approximating zero values.
}
