% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{linearRegression}
\alias{linearRegression}
\title{Generates linear regression plot for a given gene.}
\usage{
linearRegression(input, gene, assayName = NULL)
}
\arguments{
\item{input}{A list. Results from `spatialSimilarity()`. This includes the
similarity table, log-transformed pixel data, and analysis parameters.}

\item{gene}{Character. The name of the gene to visualize.}

\item{assayName}{A character string or numeric specifying the assay in the
Spatial Experiment to use. Default is \code{NULL}. If no value is supplied
for \code{assayName}, then the first assay is used as a default}
}
\value{
A ggplot2 scatter plot displaying gene expression values from two
  spatial experiments. Data points are colored as follows:
\itemize{
  \item{\strong{blue}}{Pixels classified as similar (within the fold-change threshold).}
  \item{\strong{yellow}}{Pixels with greater expression in dataset X than Y.}
  \item{\strong{red}}{Pixels with greater expression in dataset Y than X.}
  \item{\strong{grey}}{Pixels with gene expression below the threshold in both experiments.}
}
The plot includes:
\itemize{
  \item{\strong{Solid line:}}{y = x (perfect correlation).}
  \item{\strong{Dashed lines:}}{Fold-change similarity thresholds (upper and lower bounds).}
}
}
\description{
This function creates a scatter plot comparing gene expression levels between
two spatial experiments for a specified gene. It colors data points based on
similarity classification and overlays fold-change threshold lines.
}
\examples{
data(speKidney)
##### Rasterize to get pixels at matched spatial locations #####
rastKidney <- SEraster::rasterizeGeneExpression(speKidney,
               assay_name = 'counts', resolution = 0.2, fun = "mean",
               BPPARAM = BiocParallel::MulticoreParam(), square = FALSE)

s <- spatialSimilarity(list(rastKidney$A, rastKidney$C))
linearRegression(s, "Gene")

}
