% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCorrelation.R
\name{plotCorrelationGeneExp}
\alias{plotCorrelationGeneExp}
\title{plotCorrelationGeneExp}
\usage{
plotCorrelationGeneExp(speList, spatialCorrelation, geneName, assayName = NULL)
}
\arguments{
\item{speList}{\code{list} List of two SpatialExperiment objects with matched
spatial locations. The first element corresponds to the first
SpatialExperiment (`X`), and the second to the second SpatialExperiment
(`Y`).}

\item{spatialCorrelation}{\code{dataframe}: the output from running
`spatialCorrelationGeneExp` function on `speList`}

\item{geneName}{\code{character}: A character string specifiying the name of
gene (row) in both a SpatialExperiments in `speList`}

\item{assayName}{\code{character} or \code{integer} A character string or
numeric specifying the assay in the SpatialExperiment to use. Default is
\code{NULL}. If no value is supplied for \code{assayName}, then the first
assay is used as a default}
}
\value{
The output is returned as a ggplot grob. A scatterplot with the
  values of the gene in the first SpatialExperiment, i.e
  `SummarizedExperiment::assay(speList[[1]], assayName)[geneName, ]`, on the
  x-axis and values of the gene in the second SpatialExperiment on the
  y-axis. The title includes the gene name, the correlation coefficient, and
  the empirical p-value as the greater of
  `spatialCorrelation$pValuePermuteX`and `spatialCorrelation$pValuePermuteY`.
}
\description{
Function to calculate Pearson's correlation between rows from
  one SpatialExperiment dataset. To replace the analytical p-value which
  results in a high false positive rate for autocorrelated spatial patterns,
  it calculates empirical p-values from empirical null distributions
  generated from permuting the data and then smoothing to maintain the
  original degree of autocorrelation
}
\examples{

data(speKidney)

##### Rasterize to get pixels at matched spatial locations #####
rastKidney <- SEraster::rasterizeGeneExpression(speKidney,
               assay_name = 'counts', resolution = 0.2, fun = "mean",
               BPPARAM = BiocParallel::MulticoreParam(), square = FALSE)

##### Use STcompare to calculate Pearson's correlation coefficient #####
rastGexpListAB <- list(A = rastKidney$A, B = rastKidney$B)
rastGexpListAC <- list(A = rastKidney$A, C = rastKidney$C)

negCorrelation <- spatialCorrelationGeneExp(rastGexpListAB)
posCorrelation <- spatialCorrelationGeneExp(rastGexpListAC)

negCorrelation
posCorrelation

expAB <- plotCorrelationGeneExp(rastGexpListAB, negCorrelation, "Gene")
expAC <- plotCorrelationGeneExp(rastGexpListAC, posCorrelation, "Gene")

expAB
expAC
}
