% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageFunction.R
\name{spatialSimilarity}
\alias{spatialSimilarity}
\title{Computes spatial similarity of gene expression between two
SpatialExperiments.}
\usage{
spatialSimilarity(
  input,
  t1 = NULL,
  t2 = NULL,
  minQuantile = 0.05,
  foldChange = 1,
  assayName = NULL
)
}
\arguments{
\item{input}{List of two SpatialExperiment objects. The first element
corresponds to the first spatial experiment (`x`), and the second to the
second spatial experiment (`y`). Similarity is computed as the proportion
of pixels satisfying: `-b <= log2(y/x) <= b` for each gene, where `b = 1`
as default}

\item{t1}{\code{numeric}: Gene expression threshold for the first spatial
experiment (`x`). Only pixels with values greater than t1 and t2 are used
to calculate similarity score. Default is \code{NULL}. If no value is
supplied for \code{t1}, then each gene will have a threshold set as by
\code{minQuantile}.}

\item{t2}{\code{numeric}: Gene expression threshold for the second spatial
experiment (`y`). Only pixels with values greater than t1 and t2 are used
to calculate similarity score. Default is \code{NULL}. If no value is
supplied for \code{t2}, then each gene will have a threshold set as by
\code{minQuantile}.}

\item{minQuantile}{\code{numeric}: Gene expression minimum quantile threshold
to use if \code{t1} and \code{t2} are not supplied.  Default is
\code{0.05}.}

\item{foldChange}{\code{numeric}: Fold-change threshold defining similarity
in gene expression across spatial locations. Default is \code{1}.}

\item{assayName}{A character string or numeric specifying the assay in the
Spatial Experiment to use. Default is \code{NULL}. If no value is supplied
for \code{assayName}, then the first assay is used as a default}
}
\value{
A list containing: \item{similarityTable}{A data frame with the
  following columns:}
  \itemize{
    \item{\strong{gene}}{Gene name.}
    \item{\strong{percentSimilarity}}{Percentage of similar pixels for the gene.}
    \item{\strong{similarPixelID}}{List of pixel IDs classified as similar.}
    \item{\strong{dissimilarPixelIDX}}{List of pixel IDs for log2(y/x) < -foldChange.}
    \item{\strong{dissimilarPixelIDY}}{List of pixel IDs for log2(y/x) > foldChange.}
    \item{\strong{numPixelInThresh}}{Number of pixels above the threshold in both experiments.}
    \item{\strong{pixelIDInThresh}}{List of pixel IDs above the threshold.}
    \item{\strong{numPixelOutThresh}}{Number of pixels below the threshold.}
    \item{\strong{pixelIDOutThresh}}{List of pixel IDs below the threshold.}
    \item{\strong{t1}}{Threshold value for this gene in first spatial experiment.}
    \item{\strong{t2}}{Threshold value for this gene in second spatial experiment.}
  }
  \item{pixelLogTransformation}{A data frame containing log-transformed
  fold-change values for each gene.} \item{parameters}{A list of input
  parameters used in the computation.}
}
\description{
This function calculates the spatial similarity of gene expression patterns
between two SpatialExperiment objects based on thresholding and fold-change
criteria. Similarity is defined as the proportion of pixels where the fold
change in gene expression falls within a specified range.
}
