% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCorrelation.R
\name{viladomatCorrelation}
\alias{viladomatCorrelation}
\title{viladomatCorrelation}
\usage{
viladomatCorrelation(
  data,
  delta,
  maxDistPrctile,
  nPermutations,
  nThreads = 1,
  BPPARAM = NULL,
  seed = 0
)
}
\arguments{
\item{data}{\code{matrix} A N x 4 matrix of with the first column as the
values of X, the second column as the values of Y, the third column as the
x-coordinates, and the fourth column as the y-coordinates.}

\item{delta}{\code{numeric vector} Given a point in dataset X, the percentage
of neighbors which should be within the smoothing kernel. This can be a
list or a single numeric. Input to `locfit::lp` as `nn`}

\item{maxDistPrctile}{\code{numeric}: Percentile of distances between pixels
to use as max distance in when calculating variograms. At greater distances
the variogram is less precise because there are fewer pairs of points with
that distance between them. Therefore, since the goal is to minimize the
difference between the variogram of X and those of its permutations, the
variogram should be subsetted to the percentile that is more robust.}

\item{nPermutations}{\code{integer}: Number of permutations to generate to
build the empirical null distribution. This number will determine the
precision of the p-value. The smallest possible p-value is
\eqn{1 / (nPermutations + 1)}, for example about 0.0099 when
\code{nPermutations <- 100}}

\item{nThreads}{\code{integer}: Number of threads for parallelization.
Default = 1. Inputting this argument when the \code{BPPARAM} argument is
\code{NULL} would set parallel execution back-end to be
\code{BiocParallel::MulticoreParam(workers = nThreads)}. We recommend
setting this argument to be the number of cores available
(\code{parallel::detectCores(logical = FALSE)}). If \code{BPPARAM} argument
is not \code{NULL}, the \code{BPPARAM} argument would override
\code{nThreads} argument.}

\item{BPPARAM}{\code{BiocParallelParam}: Optional additional argument for
parallelization. This argument is provided for advanced users of
\code{BiocParallel} for further flexibility for setting up
parallel-execution back-end. Default is NULL. If provided, this is assumed
to be an instance of \code{BiocParallelParam}.}

\item{seed}{\code{integer}: Seed for the random number generator. Default \code{0}.}
}
\value{
The output is returned as a list.
 \itemize{
  \item{\code{deltaStarMedian}}{numeric, the median of the deltas that minimize
  the residual sum of squares across each permutation}
  \item{\code{deltaStar}}{numeric vector of length of `nPermutations`,
  the delta that minimizes the residual sum of squares for each permutation}
  \item{\code{pValueGlobal}}{numeric, empirical p-value for the Pearson's
  correlation of X and Y, computed as \eqn{(b + 1) / (B + 1)} where \eqn{b}
  is the number of null correlations whose absolute value is at least the
  absolute value of the observed correlation}
  \item{\code{nullCorGlobal}}{a B x 1 matrix, where B is `nPermutations`.
  This matrix is the correlation coefficients between the permutations and
  X that compose that null distribution used to calculate the empirical p-value}
  \item{\code{permutations}}{a N x B matrix, where B is `nPermutations`.
  Each column is the resulting values of a permutation of X}
  }
}
\description{
Function to calculate Pearson's correlation between two spatial
  datasets, X and Y. To replace the analytical p-value which results in a
  high false positive rate for autocorrelated spatial patterns, it calculates
  empirical p-values from empirical null distributions generated from
  permuting dataset X by randomly shuffling the values and then smoothing to
  maintain the original degree of autocorrelation of X
}
\examples{

data(quakes)

#remove duplicated positions
quakes_data <- quakes[!duplicated(cbind(quakes$lat, quakes$long)),]

data <- cbind(quakes_data$depth,
              quakes_data$mag,
              quakes_data$lat,
              quakes_data$long)

#sequence of deltas
delta <- seq(0.1,0.9,0.1)

# maximum distance for the variogram set at the 25\% percentile of
# the distribution of pairs of distances:
maxDistPrctile <- 0.25

#number of permutations
nPermutations <- 3

resultsPermuteX <- viladomatCorrelation(data, delta, maxDistPrctile, nPermutations)

}
\references{
adapted from: Viladomat, Júlia et al. “Assessing the significance
of global and local correlations under spatial autocorrelation: a
nonparametric approach.” Biometrics vol. 70,2 (2014): 409-18.
doi:10.1111/biom.12139
}
